// src/app/models/data.model.ts
export interface Data {
  row_ID: String;
  par_ROW_ID: String;
  lov_TYPE: String;
  lov_NAME: String;
  display_VAL: String;
  lov_VAL1: String;
  lov_VAL2: String;
  lov_VAL3: String;
  lov_VAL4: String;
  lov_VAL5: String;
  lov_VAL6: String;
  lov_VAL7: String;
  lov_VAL8: String;
  lov_VAL9: String;
  lov_VAL10: String;
  active_FLG: String;
  text_DESC: String;
  order_BY: BigInt;
  modification_NUM: BigInt;
  created: Date;
  created_BY: String;
  last_UPD: Date;
  last_UPD_BY: String;
  group_TYPE: String;
}